Fixed the inventory glitch where the inventory would not show properly when using the inventory command.
