--- BrentonFigures-MormonASCIIGame/ASCII/x64/Debug/ASCII.tlog/ASCII.lastbuildstate
@@ -1,2 +1,0 @@
-PlatformToolSet=v143:VCToolArchitecture=Native64Bit:VCToolsVersion=14.33.31629:TargetPlatformVersion=10.0.19041.0:
-Debug|x64|C:\Users\Brenton\source\repos\Game140\ASCII\ASCII\|
