--- BrentonFigures-MormonASCIIGame/ASCII/x64/Debug/ASCII.vcxproj.FileListAbsolute.txt
@@ -1,1 +1,0 @@
-C:\Users\Brenton\source\repos\Game140\ASCII\ASCII\x64\Debug\ASCII.exe
