+++ BrentonFigures-MormonASCIIGame/Figures-MormonB_ASCIIDungeon_game140_F2022/ASCII/x64/Debug/ASCII.exe.recipe
@@ -1,0 +1,11 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<Project>
+  <ProjectOutputs>
+    <ProjectOutput>
+      <FullPath>C:\Users\Brenton\source\repos\Game140\FinishedAssignments\Figures-MormonB_ASCIIDungeon_game140_F2022\Figures-MormonB_ASCIIDungeon_game140_F2022\ASCII\x64\Debug\ASCII.exe</FullPath>
+    </ProjectOutput>
+  </ProjectOutputs>
+  <ContentFiles />
+  <SatelliteDlls />
+  <NonRecipeFileRefs />
+</Project>
