+++ BrentonFigures-MormonASCIIGame/Figures-MormonB_ASCIIDungeon_game140_F2022/ASCII/x64/Debug/ASCII.vcxproj.FileListAbsolute.txt
@@ -1,0 +1,2 @@
+C:\Users\Brenton\source\repos\Game140\ASCII\ASCII\x64\Debug\ASCII.exe
+C:\Users\Brenton\source\repos\Game140\FinishedAssignments\Figures-MormonB_ASCIIDungeon_game140_F2022\Figures-MormonB_ASCIIDungeon_game140_F2022\ASCII\x64\Debug\ASCII.exe
